requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
django==1.9
django-braces==1.8.1
django-ckeditor==5.0.2
django-debug-toolbar==1.4
django-grappelli==2.7.2
django-toolbelt==0.0.1
gunicorn==19.4.5
pbr==1.8.1
Pillow==3.0.0
psycopg2==2.6.1
six==1.10.0
sqlparse==0.1.18
static3==0.6.1
stevedore==1.9.0
Unipath==1.1
virtualenv==14.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
